prototype/tests/non-secure-context-dynamic-import.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Async local storage: should not work in non-secure contexts when included via import()</title>

<script src="https://w3c-test.org/resources/testharness.js"></script>
<script src="https://w3c-test.org/resources/testharnessreport.js"></script>

<script>
"use strict";

test(() => {
  assert_false(window.isSecureContext, "This test must run in a non-secure context");
}, "Prerequisite check");

promise_test(t => {
  return promise_rejects(t, "SecurityError", import("../implementation.mjs"));
});
</script>
